+++ demos/chinook/src/main/java/is/codion/framework/demos/chinook/ui/CustomerTablePanel.java
@@ -28,7 +28,7 @@ public final class CustomerTablePanel extends EntityTablePanel {
 
   public CustomerTablePanel(SwingEntityTableModel tableModel) {
     super(tableModel);
-    control(ControlCode.PRINT).set(Control.builder(this::viewCustomerReport)
+    control(TableControl.PRINT).set(Control.builder(this::viewCustomerReport)
             .name(BUNDLE.getString("customer_report"))
             .smallIcon(FrameworkIcons.instance().print())
             .enabled(tableModel().selectionModel().selectionNotEmpty())