+++ demos/world/src/main/java/is/codion/framework/demos/world/ui/CountryTablePanel.java
@@ -23,7 +23,7 @@ final class CountryTablePanel extends EntityTablePanel {
   CountryTablePanel(SwingEntityTableModel tableModel) {
     super(tableModel);
     editableAttributes().remove(Country.CAPITAL_FK);
-    control(ControlCode.PRINT).set(Control.builder(this::viewCountryReport)
+    control(TableControl.PRINT).set(Control.builder(this::viewCountryReport)
             .name("Country report")
             .enabled(tableModel.selectionModel().selectionNotEmpty())
             .smallIcon(FrameworkIcons.instance().print())